+++ src/php/DataSift/Storyplayer/PlayerLib/Story/Result.php
@@ -84,7 +84,7 @@ class Story_Result extends PhaseGroup_Result
     public function __construct(Story $story)
     {
         // initialise our parent first
-        parent::__construct($story->getCategory() . ' > ' . $story->getGroupAsString() . ' > ' . $story->getName());
+        parent::__construct($story->getNameForConsole());
 
         // remember the story we are reporting on
         $this->story = $story;