+++ CHANGELOG.md
@@ -25,6 +25,7 @@ In this release, I've made some important improvements to Storyplayer's `runtime
 
 * Story templates should now extend `Storyplayer\SPv2\Stories\StoryTemplate`. The original `DataSift\Storyplayer\PlayerLib\StoryTemplate` base class is still there, for backwards-compatibility.
 * New stories should now be created from `Storyplayer\SPv2\Stories\BuildStory`.
+  * These stories will show their filename in output, rather than the old category / group / called triad.
 
 ### Refactor